Fixed compilation errors & completed method
As of now, this program does the following:
- Initializes a database which reads from our inventory_team1.csv file
- Creates a BuyerEvent object using parameters from buyer_event.csv file
- Updates the quantity of a product given it's product_id
- Updates the database with the new quantity of the product

Note: If you run the program, I also printed out a single product BEFORE the simulation is run and again AFTER, so that you can visually see the changes to quantity. We may want to remove that before merging to the master branch.